docs: neonctl 2.25.0, includes bootstrap and bucket object put
