+++ src/components/pages/doc/cli-reference/cli-command-index/groups.js
@@ -14,6 +14,7 @@ const GROUPS = [
 const GROUP_OF = {
   auth: 'setup',
   init: 'setup',
+  bootstrap: 'setup',
   link: 'setup',
   checkout: 'setup',
   'set-context': 'setup',
